fix circular reference in shared and indexer
